+++ src/lib.rs
@@ -70,7 +70,7 @@ pub use gcode::{
     stream::{FileStreamReader, GcodeStreamReader, PausableStream, StringStreamReader},
 };
 pub use ui::{SettingsDialog, Setting, SettingValue, SettingsCategory, KeyboardShortcut, SettingsPersistence, FirmwareSettingsIntegration, DeviceConsoleManager, DeviceMessageType, ConsoleListener, GcodeEditor, GcodeLine, Token, TokenType};
-pub use utils::{FileEncoding, FileReadStats, FileValidation, GcodeFileReader, RecentFileEntry, RecentFilesManager, BoundingBox, FeedRateStats, FileProcessingPipeline, FileStatistics, ProcessedFile, SpindleStats};
+pub use utils::{FileEncoding, FileReadStats, FileValidation, GcodeFileReader, RecentFileEntry, RecentFilesManager, BoundingBox, FeedRateStats, FileProcessingPipeline, FileStatistics, ProcessedFile, SpindleStats, DropEvent, DropFileType, DropIndicatorState, DropTarget, DropZone, ExportOptions, FileExporter, FileFormat};
 
 /// Library version
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");